% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datetime.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetime from filenames (YYYYMMDD_HHMMSS or similar)}
\usage{
parse_datetime(filenames)
}
\arguments{
\item{filenames}{A character, ususaly the name of the file, including the date and time of recording}
}
\value{
a POSIXct object/string
}
\description{
Parse datetime from filenames (YYYYMMDD_HHMMSS or similar)
}
\examples{
# example code
filenames = c("path/to/20240719_112000.WAV", "another/path/20240722_124000.WAV")
datetimes <- parse_datetime(filenames)
print(datetimes)

}
