% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictions}
\alias{predictions}
\title{Exemplary predictions of Orthoptera on audio recordings}
\format{
A data frame with 137968 rows and 34 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
\usage{
predictions
}
\description{
Exemplary predictions of Orthoptera on audio recordings
}
\keyword{datasets}
