% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_predictions.R
\name{reshape_predictions}
\alias{reshape_predictions}
\title{Reshape ML wide predictions to long tidy format}
\usage{
reshape_predictions(df, tz = "UTC")
}
\arguments{
\item{df}{data.frame with filename, offset, and one column per species (confidence values)}

\item{tz}{character timezone (default "UTC")}
}
\value{
data.table with columns filename, offset, species, conf, timestamp, Plot_ID
}
\description{
Reshape ML wide predictions to long tidy format
}
