% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_ranges.R
\name{get_gbif_ranges}
\alias{get_gbif_ranges}
\title{Build GBIF-based species ranges at selected resolution}
\usage{
get_gbif_ranges(
  species,
  res = 10,
  limit = 200,
  region = "world",
  collect_citations = FALSE
)
}
\arguments{
\item{species}{Character vector of species names, or a column from a data.frame. Can either be formated like
c("Bicolorana bicolor", "Chorthippus albomarginatus") or c("Bicolorana.bicolor","Chorthippus.albomarginatus") to match with model output nomeclature.}

\item{res}{Numeric. Grid resolution in kilometers. Default is 10.}

\item{limit}{Integer. Maximum number of GBIF records per species. Default is 200.}

\item{region}{Character. One of "world", "europe", "asia", "americas".
Defines the bounding box used for the GBIF query. Default is "world".}

\item{collect_citations}{Logical. If TRUE, dataset citations are retrieved
and attached as an attribute of the returned sf object.}
}
\value{
An \code{terra} raster stack object with a \code{species} occurence raster stacked. If \code{collect_citations=TRUE},
a \code{citations} attribute is attached (data.frame with datasetKey, title, citation).
}
\description{
Retrieves occurrence data from GBIF via \pkg{rgbif} and aggregates them
into a gridded species range map at the specified resolution.
}
\details{
The result is an \code{terra} raster stack suitable for spatial plausibility checks.
A citation for the used datsets can also be exported. Please cite the data accordingly.
}
\examples{
\dontrun{
sp_vec <- c("Tettigonia viridissima", "Barbitistes serricauda")

# Retrieve ranges for Europe only
ranges_eu <- get_gbif_ranges(
  species = species_vec,
  res = 50,
  limit = 100,
  region = "europe",
  collect_citations = TRUE)

# Plot result
plot(ranges_eu)

# plot citations

}

}
