% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datetime_and_location.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetime from filename (YYYYMMDD_HHMMSS or similar)}
\usage{
parse_datetime(df, filename_col = "filename")
}
\arguments{
\item{filename}{A character, ususaly the name of the file, including the date and time of recording}

\item{tz}{time zone (default "UTC")}
}
\value{
a POSIXct object/string
}
\description{
Parse datetime from filename (YYYYMMDD_HHMMSS or similar)
}
\examples{
# example code
predictions <- data.frame(filename = c("path/to/20240719_112000.WAV", "another/path/20240722_124000.WAV"))
datetimes <- parse_datetime(predictions)
print(datetimes)

}
